### 2024-04-12 #6613 @editorialbot
[PRE REVIEW]: tRIBS v5.2: A multi-resolution, parallel platform for tributary hydrology in forest applications
**Submitting author:** @WrenRaming (<a href="http://orcid.org/0000-0003-2204-4621">L. Wren Raming</a>)
**Repository:** https://github.com/tribshms/tRIBS
**Branch with paper.md** (empty if default branch):
**Version:** 5.2.0
