### 2024-04-15 #6613 @kthyng
@WrenRaming Your [license file](https://github.com/tribshms/tRIBS/blob/main/LICENSE.md) has text in it beyond the text of the actual license itself. Can you change this to be the plain MIT license file "LICENSE" with no modification to the text and not a markdown file?
