### 2024-04-15 4705541
Updated LICENSE file to text and removed extraneous information
