### 2024-04-15 #6613 @WrenRaming
Hi @kthyng I have updated the LICENSE so that only the license information remains and I have converted into a text file.  In regards to installing as package I think that shouldn't be a problem. I will start looking into that.
