+++ README.md
@@ -9,7 +9,7 @@
 # TIN-based Real-time Integrated Basin Simulator: Version 5.2
 This repository contains source code for the fully distributed hydrological model: TIN-based Real-time Integrated Basin Simulator (tRIBS). Details on running the model and its applications can be found [here](https://tribshms.readthedocs.io/en/latest/).
 
-Version and licensing information can be found in [LICENSE.md](./LICENSE.md).
+Licensing information can be found in [LICENSE.md](./LICENSE.md).
 
 ## Installation and Compilation
 tRIBS is an object-oriented based code written in c++. In order to run tRIBS you must download and compile the source code from this GitHub repository.
