### 2024-04-17 edfd601
Fixed if/else statement for release build and update CMakeList.txt for packaging and passing commands for parallel or serial build with different names for both
