+++ .github/workflows/compile_and_test_macos_intel.yml
@@ -11,7 +11,7 @@ on:
 
 jobs:
   build_and_test:
-    runs-on: macos-latest
+    runs-on: macos-13 #Intel Chip
     steps:
       - name: Checkout Repository
         uses: actions/checkout@v2
@@ -26,14 +26,14 @@ jobs:
 
       # Build Parallel Version of tRIBS
       - name: Create build directory and run CMake for parallel version
-        run: cmake -S . -B cmake-build_parallel -DCMAKE_BUILD_TYPE=Release -Dparallel=ON -Dpackaging=ON -Dos=macOS/Silicon
+        run: cmake -S . -B cmake-build_parallel -DCMAKE_BUILD_TYPE=Release -Dparallel=ON -Dpackaging=ON -Dos=macOS/Intel
 
       - name: Build Project Parallel
         run: cmake --build cmake-build_parallel --target all
 
       # Build Serial Version of tRIBS
       - name: Create build directory and run CMake for serial version
-        run: cmake -S . -B cmake-build_serial -DCMAKE_BUILD_TYPE=Release -Dparallel=OFF -Dpackaging=ON -Dos=macOS/Silicon
+        run: cmake -S . -B cmake-build_serial -DCMAKE_BUILD_TYPE=Release -Dparallel=OFF -Dpackaging=ON -Dos=macOS/Intel
 
       - name: Build Project Serial
         run: cmake --build cmake-build_serial --target all
@@ -46,5 +46,5 @@ jobs:
       - name: Package artifacts
         uses: actions/upload-artifact@v2
         with:
-          name: tRIBS-STGZ-macos-Silicon
-          path: "packaging/macOS/Silicon/TIN-based Real-time Integrated Basin Simulator-5.2.0.sh"
+          name: tRIBS-STGZ-macos-Intel
+          path: "packaging/macOS/Intel/TIN-based Real-time Integrated Basin Simulator-5.2.0.sh"
