+++ .github/workflows/compile_and_test_linux.yml
@@ -11,16 +11,7 @@ on:
 
 jobs:
   build_and_test:
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest]
-        compiler: [g++-11]
-        include:
-          - os: ubuntu-latest
-            compiler: g++-11
-            target: Linux
-
+    runs-on: ubuntu-latest
     steps:
       - name: Checkout Repository
         uses: actions/checkout@v2
@@ -33,8 +24,27 @@ jobs:
       - name: Set CMAKE_PREFIX_PATH
         run: export CMAKE_PREFIX_PATH=/usr/lib/openmpi:$CMAKE_PREFIX_PATH
 
-      - name: Create build directory and run CMake
-        run: cmake -S . -B build -DCMAKE_BUILD_TYPE=Debug
+      # Build Parallel Version of tRIBS
+      - name: Create build directory and run CMake for parallel version
+        run: cmake -S . -B cmake-build_parallel -DCMAKE_BUILD_TYPE=Release -Dparallel=ON -Dpackaging=ON -Dos=Linux
 
-      - name: Build Project
-        run: cmake --build build --target all
+      - name: Build Project Parallel
+        run: cmake --build cmake-build_parallel --target all
+
+      # Build Serial Version of tRIBS
+      - name: Create build directory and run CMake for serial version
+        run: cmake -S . -B cmake-build_serial -DCMAKE_BUILD_TYPE=Release -Dparallel=OFF -Dpackaging=ON -Dos=Linux
+
+      - name: Build Project Serial
+        run: cmake --build cmake-build_serial --target all
+
+      # Run CPack
+      - name: Package tRIBS
+        run: cpack --config  packaging/utilities/MultiCPackConfig.cmake
+
+      # Upload STGZ package
+      - name: Package artifacts
+        uses: actions/upload-artifact@v2
+        with:
+          name: tRIBS-STGZ-linux
+          path: "packaging/Linux/TIN-based Real-time Integrated Basin Simulator-5.2.0.sh"
