+++ .github/workflows/compile_and_test_linux.yml
@@ -14,7 +14,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout Repository
-        uses: actions/checkout@v2
+        uses: actions/checkout@v4
 
       # Install MPI Development Files
       - name: Install MPI Development Files
@@ -44,7 +44,7 @@ jobs:
 
       # Upload STGZ package
       - name: Package artifacts
-        uses: actions/upload-artifact@v2
+        uses: actions/upload-artifact@v4
         with:
           name: tRIBS-STGZ-linux
           path: "packaging/Linux/TIN-based Real-time Integrated Basin Simulator-5.2.0.sh"
