+++ .github/workflows/compile_and_test_macos.yml
@@ -14,7 +14,7 @@ jobs:
     runs-on: macos-latest
     steps:
       - name: Checkout Repository
-        uses: actions/checkout@v2
+        uses: actions/checkout@v4
 
       # Install MPI Development Files
       - name: Install MPI Development Files
@@ -44,7 +44,7 @@ jobs:
 
       # Upload STGZ package
       - name: Package artifacts
-        uses: actions/upload-artifact@v2
+        uses: actions/upload-artifact@v4
         with:
           name: tRIBS-STGZ-macos-Silicon
           path: "packaging/macOS/Silicon/TIN-based Real-time Integrated Basin Simulator-5.2.0.sh"
