+++ .github/workflows/compile_and_test_macos.yml
@@ -11,7 +11,7 @@ on:
 
 jobs:
   build_and_test:
-    runs-on: macos-latest
+    runs-on: macos-14
     steps:
       - name: Checkout Repository
         uses: actions/checkout@v4
@@ -21,8 +21,8 @@ jobs:
         run: brew install open-mpi
 
       # Set CMAKE_PREFIX_PATH
-      - name: Set CMAKE_PREFIX_PATH
-        run: export CMAKE_PREFIX_PATH=/usr/local/opt/open-mpi:$CMAKE_PREFIX_PATH
+      #- name: Set CMAKE_PREFIX_PATH
+      #  run: export CMAKE_PREFIX_PATH=/usr/local/opt/open-mpi:$CMAKE_PREFIX_PATH
 
       # Build Parallel Version of tRIBS
       - name: Create build directory and run CMake for parallel version
