+++ .github/workflows/compile_and_test_macos_intel.yml
@@ -21,8 +21,8 @@ jobs:
         run: brew install open-mpi
 
       # Set CMAKE_PREFIX_PATH
-      - name: Set CMAKE_PREFIX_PATH
-        run: export CMAKE_PREFIX_PATH=/usr/local/opt/open-mpi:$CMAKE_PREFIX_PATH
+      #- name: Set CMAKE_PREFIX_PATH
+      #  run: export CMAKE_PREFIX_PATH=/usr/local/opt/open-mpi:$CMAKE_PREFIX_PATH
 
       # Build Parallel Version of tRIBS
       - name: Create build directory and run CMake for parallel version
