+++ README.md
@@ -7,11 +7,11 @@
 ![](https://img.shields.io/readthedocs/tribshms)
 
 # TIN-based Real-time Integrated Basin Simulator: Version 5.2
-This repository contains source code for the fully distributed hydrological model: TIN-based Real-time Integrated Basin Simulator (tRIBS). We provide extensive documention of the model [here](https://tribshms.readthedocs.io/en/latest/). Our documentation also contains useful information about working with the tRIBS code base through [GitHub](file:///Users/lraming/Documents/Repos/Forked/rtdocs/docs/_build/html/man/Using%20GitHub.html) and how to [contribute](file:///Users/lraming/Documents/Repos/Forked/rtdocs/docs/_build/html/man/Contributing.html). Finally we also provide [templates](file:///Users/lraming/Documents/Repos/Forked/rtdocs/docs/_build/html/man/Executables.html) and [benchmarks](file:///Users/lraming/Documents/Repos/Forked/rtdocs/docs/_build/html/man/Benchmarks.html) with instruction on how to utilize these resources in the documentation. 
+This repository contains source code for the fully distributed hydrological model: TIN-based Real-time Integrated Basin Simulator (tRIBS). We provide extensive documention of the model [here](https://tribshms.readthedocs.io/en/latest/). Our documentation also contains useful information about working with the tRIBS code base through [GitHub](https://tribshms.readthedocs.io/en/latest/man/Using%20GitHub.html) and how to [contribute](https://tribshms.readthedocs.io/en/latest/man/Contributing.html). Finally we also provide [templates](https://tribshms.readthedocs.io/en/latest/man/Templates.html) and [benchmarks](https://tribshms.readthedocs.io/en/latest/man/Benchmarks.html) with instruction on how to utilize these resources in the documentation. 
 
 Licensing information can be found in [LICENSE.txt](./LICENSE.txt).
 
-Release notes are provided [here](file:///Users/lraming/Documents/Repos/Forked/rtdocs/docs/_build/html/man/Release%20Notes.html).
+Release notes are provided [here](https://tribshms.readthedocs.io/en/latest/man/Release%20Notes.html#).
 
 ## Installation 
 We provide three options for accessing the tRIBS model:
