### 2024-04-22 1ded824
Updated docker file so mpirun is not called under the root user, and fixed mpi issue where MPI_INTEGER should be MPI_INT for C/C++ code. This allows tRIBS to run on Alpine without needing to bring in fortran
