### 2024-04-22 #6613 @WrenRaming
Hi @kthyng, I have put in a bit of effort figuring how best to package the software. The best solution I found was to use CPack to create a self extracting tar gzip. These can be found on the [executables page](https://tribshms.readthedocs.io/en/latest/man/Executables.html#executables) in our documentation —though these will only be compatible with the latest macOS and Ubuntu systems. Alternatively, I would argue the Docker image also provides a good alternative as documented [here](https://tribshms.readthedocs.io/en/latest/man/Docker.html). I am not sure if this is exactly what you are looking for, but any additional suggestions or examples are more than welcomed. Noted on the waitlist, thanks for your tireless effort!
