### 2024-04-23 #6613 @kthyng
@WrenRaming Regarding packaging, we want to have a reasonable way for reviewers and your future users to install your software using the community standards for the language. Docker is a good alternative when nothing else will work, but we also more basic approaches so that the software is accessible for developers. It sounds like you have done the necessary thinking for this process, thank you! The reviewers will probably have more and better input than me.
