### 2024-05-11 b4b276a
A few minor patches related to vegetation fraction and snow dynamics--also updated CHANGELOG.md with more extensive release notes, including current updates.
