+++ README.md
@@ -6,7 +6,7 @@
 
 ![](https://img.shields.io/readthedocs/tribshms)
 
-# TIN-based Real-time Integrated Basin Simulator: Version 5.2
+# TIN-based Real-time Integrated Basin Simulator: Version 5.2.0
 This repository contains source code for the fully distributed hydrological model: TIN-based Real-time Integrated Basin Simulator (tRIBS). We provide extensive documention of the model [here](https://tribshms.readthedocs.io/en/latest/). Our documentation also contains useful information about working with the tRIBS code base through [GitHub](https://tribshms.readthedocs.io/en/latest/man/Using%20GitHub.html) and how to [contribute](https://tribshms.readthedocs.io/en/latest/man/Contributing.html). Finally we also provide [templates](https://tribshms.readthedocs.io/en/latest/man/Templates.html) and [benchmarks](https://tribshms.readthedocs.io/en/latest/man/Benchmarks.html) with instruction on how to utilize these resources in the documentation. 
 
 Licensing information can be found in [LICENSE.txt](./LICENSE.txt).
@@ -19,3 +19,6 @@ We provide three options for accessing the tRIBS model:
 1) Compiled executables for latest [macOS and Ubuntu](https://tribshms.readthedocs.io/en/latest/man/Executables.html#executables) systems
 2) [Docker](https://tribshms.readthedocs.io/en/latest/man/Docker.html)
 3) [CMake build system](https://tribshms.readthedocs.io/en/latest/man/Model_Execution.html#cmake)
+
+## Release/Version Notes
+tRIBS uses semantic versioning. We record updates of major, minor, and patch versions [here](./doc/md/CHANGELOG.md).
