### 2024-05-13 #6613 @alessandroamaranto
Sure, I will most likely complete the review within the next 10 days.
Sorry, I am a little late but I will catch up :)

On Mon, May 13, 2024 at 2:53 PM Taher Chegini ***@***.***>
wrote:

> Hi @alessandroamaranto <https://github.com/alessandroamaranto>! Would you
> like to review this submission to the Journal for Open Source Software? Our
> reviews are checklist-driven and openly conducted on GitHub over a timeline
> of 4–6 weeks. Because the process is much more iterative and interactive
> than a traditional paper review, we would ask you to start within the next
> 2 weeks. Here are reviewer guidelines for reference:
> joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> Thanks for your consideration.
>
