### 2024-05-13 #6747 @cheginit
@WrenRaming, there's an issue with generating PDF from the paper. You need to remove or rename the [`JATS`](https://github.com/tribshms/tRIBS/blob/main/paper/paper.jats) file, since our `editorbot` should generate that file.
