### 2024-05-13 #6747 @WrenRaming
Hi @cheginit, I've removed the JATS file as requested.
