+++ src/tHydro/tEvapoTrans.cpp
@@ -224,11 +224,7 @@ void tEvapoTrans::SetEvapTVariables(tInputFile &infile, tHydroModel *hydro)
 		soilPtr = hydro->soilPtr;
 		initializeVariables();
 		SetSunVariables();
-		
-		tMeshListIter<tCNode> nodeIter(gridPtr->getNodeList());
-		BasAltitude  = (nodeIter.FirstP()->getZ());
-		BasAltitude += (nodeIter.LastP()->getZ());
-		BasAltitude /= 2.0;
+
 		
 		if (gFluxOption != 1 && gFluxOption != 2) {
 			Cout<<"\nGround Heat Flux Option "<< gFluxOption;
@@ -1039,13 +1035,9 @@ void tEvapoTrans::ComputeETComponents(tIntercept *Intercept, tCNode *cNode,
 		// Set Coefficients
 		if (luOption == 1) {
 			newLUGridData(cNode);
-			if (gFluxOption == 1 || gFluxOption == 2) {
-				// Giuseppe 2016 - Begin changes to allow reading soil properties from grids
-                //			        coeffKs = soilPtr->getSoilProp(10);
-                //				coeffCs = soilPtr->getSoilProp(11);
+			if (gFluxOption == 1 || gFluxOption == 2) {;
                 coeffKs = cNode->getVolHeatCond();
                 coeffCs = cNode->getSoilHeatCap();
-				// Giuseppe 2016 - End changes to allow reading soil properties from grids
             }
 		}
 		else{
@@ -1552,7 +1544,8 @@ void tEvapoTrans::SetSunVariables()
 	double Wo = 1367.0;
 	double pi = 4.0*atan(1.0);
 	double CIRC[12] = {.07, .10, .12, .14, .16, .19, .23, .20, .16, .12, .08, .07};
-	
+
+
 	dgmt  = gmt;
 	longSM = 15.0*abs(gmt);
 	longM = fabs(longitude);
@@ -2246,7 +2239,9 @@ double tEvapoTrans::energyBalance(tCNode* cNode)
 	double Tg, eps;
 	double cosi,v;
 	double Isw; // Following were removed as shadows WR: Ic,Ics,Id,Ids,Is
-	SunHour=0;
+
+
+    SunHour=0;
 	Ic=Ics=Id=Ids=Is=Isw=0.0;
     elevation = cNode->getZ(); //SMM 10172008
 	HeatTransferProperties( cNode ); 
@@ -2317,7 +2312,7 @@ double tEvapoTrans::energyBalance(tCNode* cNode)
 	// Compute resistances for turbulent fluxes
 	Rah  = aeroResist();
 	Rstm = stomResist();
-	
+
 	Tg = Tso;  // Surface toK from the previous time step
 	
 	// Iteratively find the ground temperature
@@ -5206,8 +5201,7 @@ void tEvapoTrans::writeRestart(fstream & rStr) const
   BinaryWrite(rStr, Id);
   BinaryWrite(rStr, Ids);
   BinaryWrite(rStr, vPressC);
-  BinaryWrite(rStr, Epot);
-  BinaryWrite(rStr, BasAltitude);
+  BinaryWrite(rStr, Epot);;
 
   BinaryWrite(rStr, numStations);
   BinaryWrite(rStr, arraySize);
@@ -5323,7 +5317,6 @@ void tEvapoTrans::writeRestart(fstream & rStr) const
   BinaryWrite(rStr, ha3375);
   BinaryWrite(rStr, tempLapseRate);
   BinaryWrite(rStr, SunHour);
-  BinaryWrite(rStr, BasAltitude);
[…]
