### 2024-05-23 b583129
Updated skycover_flag so it's properly initialized.
