### 2024-05-30 bb2c227
Updated snowpack code so no implicit type version is required in variable defintions--also began to implement development of black-box tests for happy jack.
