### 2024-05-31 d7bc100
updated test_inout.py to account for differences between floats
