### 2024-05-31 4efa0d4
Added important comment about inShortWaveCan coming from tSnowIntercept.cpp which was removed. I almost forgot.
