### 2024-06-03 7c75adf
Updated tests for happy jack benchmark, commented out inShortWaveCan in tSnow pack, and fixed implicit type conversion.
