+++ testing/bb_auto/happy_jack/test_inout.py
@@ -4,82 +4,159 @@
 
 # pytest --binary /path/to/your/binary/executable --input /path/to/input/file
 
-import sys
-import os
 import numpy as np
-import pytest
 
-extra_path = f"{os.environ['HOME']}/Documents/Repos/Forked/pytRIBS"
-if extra_path not in sys.path:
-    sys.path.append(extra_path)
+def test_match_precip_input_to_output(setup_data):
+    pixel, _, input_precip, m, _ = setup_data
 
-from pytRIBS.classes import Model as model
-from pytRIBS.classes import Results as results
+    p_in = m.read_precip_station(input_precip)
+    # Drop the last hour from precipitation data
+    # this would automatically fail the test if not in place because last hour is not recorded.
+    # This should be considered a bug.
+    p_in.drop(p_in.index[-1], inplace=True)
+
+    # Compare precipitation inputs and outputs
+    pdiff = p_in.R.values - pixel.Rain_mm_h.values
+
+    print("\nTesting Precip Match ")
+    assert not np.any(pdiff), "Difference found in precipitation forcing"
+
+def test_match_PA_input_to_output(setup_data):
+    pixel, input_met, _, m, _ = setup_data
+    # Read meteorological input
+    met_in = m.read_met_station(input_met)
+    # Drop the last hour from meteorological data
+    met_in.drop(met_in.index[-1], inplace=True)
+
+    tol = 0.01
+    invar = 'PA'
+    pvar = 'Press_Pa'
+
+    var_dif = np.mean(met_in[invar].values - pixel[pvar].values)
+
+    print("\nTesting Pressure Match ")
+    assert var_dif <= tol, f"Difference found in variables {invar} and {pvar} at time {pixel.Time_hr[(met_in[invar] != pixel[pvar]).idxmax()]}"
+def test_match_RH_input_to_output(setup_data):
+    pixel, input_met, _, m, _ = setup_data
+    # Read meteorological input
+    met_in = m.read_met_station(input_met)
+    # Drop the last hour from meteorological data
+    met_in.drop(met_in.index[-1], inplace=True)
+
+    tol = 0.01
+    invar = 'RH'
+    pvar = 'RelHum_[]'
+
+    var_dif = np.mean(met_in[invar].values - pixel[pvar].values)
+
+    print("\nTesting Relative Humidity Match ")
+    assert var_dif <= tol, f"Difference found in variables {invar} and {pvar} at time {pixel.Time_hr[(met_in[invar] != pixel[pvar]).idxmax()]}"
+
+def test_match_XC_input_to_output(setup_data):
+    pixel, input_met, _, m, _ = setup_data
+    # Read meteorological input
+    met_in = m.read_met_station(input_met)
+    # Drop the last hour from meteorological data
+    met_in.drop(met_in.index[-1], inplace=True)
+
+    tol = 0.01
+    tribs_nodata = 9999.99
+    invar = 'XC'
+    pvar = 'SkyCov_[]'
+
+    var_dif = np.mean(met_in[invar].values - pixel[pvar].values)
+
+    print("\nTesting Sky Cover Match")
+
+    # SkyCov_[] should all be less than 10 if estimated by tRIBS, which happens when supplied with nodata
+    if all(met_in['XC'] == tribs_nodata):
+        assert all(pixel[pvar]<=10.0)
+    else:
+        assert var_dif <= tol, f"Difference found in variables {invar} and {pvar} at time {pixel.Time_hr[(met_in[invar] != pixel[pvar]).idxmax()]}"
+def test_match_US_input_to_output(setup_data):
+    pixel, input_met, _, m, _ = setup_data
+    # Read meteorological input
[…]
