### 2024-06-05 cb88597
Working on black box tests + increased precision of time in .qout file.
