+++ testing/bb_auto/happy_jack/test_happy_jack.py
@@ -1,13 +1,9 @@
-# This test confirms if the inputs for a provided model match the outputs.
-# Currently this only setup for a single element case, but will need to be updated to address more complex multi-element
-# models.
-
-# pytest --binary /path/to/your/binary/executable --input /path/to/input/file
-
 import numpy as np
+import pandas as pd
+
 
 def test_match_precip_input_to_output(setup_data):
-    pixel, _, input_precip, m, _ = setup_data
+    pixel, _, input_precip, m, _, _ = setup_data
 
     p_in = m.read_precip_station(input_precip)
     # Drop the last hour from precipitation data
@@ -21,8 +17,9 @@ def test_match_precip_input_to_output(setup_data):
     print("\nTesting Precip Match ")
     assert not np.any(pdiff), "Difference found in precipitation forcing"
 
+
 def test_match_PA_input_to_output(setup_data):
-    pixel, input_met, _, m, _ = setup_data
+    pixel, input_met, _, m, _, _ = setup_data
     # Read meteorological input
     met_in = m.read_met_station(input_met)
     # Drop the last hour from meteorological data
@@ -36,8 +33,10 @@ def test_match_PA_input_to_output(setup_data):
 
     print("\nTesting Pressure Match ")
     assert var_dif <= tol, f"Difference found in variables {invar} and {pvar} at time {pixel.Time_hr[(met_in[invar] != pixel[pvar]).idxmax()]}"
+
+
 def test_match_RH_input_to_output(setup_data):
-    pixel, input_met, _, m, _ = setup_data
+    pixel, input_met, _, m, _, _ = setup_data
     # Read meteorological input
     met_in = m.read_met_station(input_met)
     # Drop the last hour from meteorological data
@@ -52,8 +51,9 @@ def test_match_RH_input_to_output(setup_data):
     print("\nTesting Relative Humidity Match ")
     assert var_dif <= tol, f"Difference found in variables {invar} and {pvar} at time {pixel.Time_hr[(met_in[invar] != pixel[pvar]).idxmax()]}"
 
+
 def test_match_XC_input_to_output(setup_data):
-    pixel, input_met, _, m, _ = setup_data
+    pixel, input_met, _, m, _, _ = setup_data
     # Read meteorological input
     met_in = m.read_met_station(input_met)
     # Drop the last hour from meteorological data
@@ -70,11 +70,13 @@ def test_match_XC_input_to_output(setup_data):
 
     # SkyCov_[] should all be less than 10 if estimated by tRIBS, which happens when supplied with nodata
     if all(met_in['XC'] == tribs_nodata):
-        assert all(pixel[pvar]<=10.0)
+        assert all(pixel[pvar] <= 10.0)
     else:
         assert var_dif <= tol, f"Difference found in variables {invar} and {pvar} at time {pixel.Time_hr[(met_in[invar] != pixel[pvar]).idxmax()]}"
+
+
 def test_match_US_input_to_output(setup_data):
-    pixel, input_met, _, m, _ = setup_data
+    pixel, input_met, _, m, _, _ = setup_data
     # Read meteorological input
     met_in = m.read_met_station(input_met)
     # Drop the last hour from meteorological data
@@ -85,8 +87,9 @@ def test_match_US_input_to_output(setup_data):
     print("\nTesting Wind Speed Match ")
[…]
