### 2024-06-05 a4923fd
Refining black box tests
