+++ testing/bb_auto/big_spring/test_big_spring.py
@@ -1,3 +1,7 @@
+import numpy as np
+import pandas as pd
+
+
 def test_parallel_serial_match(setup_data):
     print("\nTesting congruence between parallel and serial simulations")
     parallel_result, serial_result, _ = setup_data
@@ -29,20 +33,63 @@ def test_parallel_waterbalance(setup_data):
     # so sat and unsat fluxes = 0, also currently there is no average canopy storage variables so this portion of the
     # storage is ignored.
 
-    unsat_mm = (mrf['MSMU'].iloc[len(mrf)-1] * mrf['MDGW'].iloc[len(mrf)-1] - mrf['MSMU'].iloc[0] * mrf['MDGW'].iloc[0]) * porosity
-    sat_mm = (mrf['MDGW'].iloc[0] - mrf['MDGW'].iloc[0]) * porosity
-    can_swe_mm = 10 * (mrf['AvInSn'].iloc[len(mrf)-1] - mrf['AvInSn'].iloc[0])
-    swe_mm = 10 * (mrf['AvSWE'].iloc[len(mrf)-1] - mrf['AvSWE'].iloc[0])
+    unsat_mm = (mrf['MSMU'].iloc[-1] * mrf['MDGW'].iloc[-1] - mrf['MSMU'].iloc[0] *
+                mrf['MDGW'].iloc[0]) * porosity
+    sat_mm = (mrf['MDGW'].iloc[0] - mrf['MDGW'].iloc[-1]) * porosity
+    can_swe_mm = 10 * (mrf['AvInSn'].iloc[-1] - mrf['AvInSn'].iloc[0])
+    swe_mm = 10 * (mrf['AvSWE'].iloc[-1] - mrf['AvSWE'].iloc[0])
     p_mm = mrf['MAP'].sum()
-    et_mm = mrf['MET'].sum() - 10*(mrf['AvSnSub'].sum() + mrf['AvSnEvap'].sum() + mrf['AvInSu'].sum())
+    et_mm = mrf['MET'].sum() - 10 * (mrf['AvSnSub'].sum() + mrf['AvSnEvap'].sum() + mrf['AvInSu'].sum())
     q_mm = mrf['Srf'].sum() * 3600 * 1000 / drainage_area,
 
     del_s = unsat_mm + sat_mm + can_swe_mm + swe_mm
     loss = et_mm + q_mm
-    test = (p_mm - del_s - loss)/years # mm/yr
+    test = (p_mm - del_s - loss) / years  # mm/yr
 
     # test avg. annual value
     print(f"Test metric (P - Loss - delS) = {test} mm/yr")
 
-    assert test < wb_threshold
+    assert abs(test) < wb_threshold
 
+
+def test_parallel_efficiency_match(setup_data):
+    mrf, _, r_par = setup_data
+
+    ref_kge = 1 - np.sqrt(2)
+
+    print(f"\nTesting Model Efficiency.\nThreshold = {ref_kge} from Koben et al. 2011")
+
+    # model results
+    qout = r_par.get_qout_results()
+    qout.set_index('Time', inplace=True)
+    qout_hour = qout.resample('h').mean()
+
+    # need to be updated so it's more portable
+    obs_data = '/Users/wr/Desktop/big_spring/data/obs/Discharge_MS-1.csv'
+
+    q_obs = pd.read_csv(obs_data, skiprows=24)
+
+    # convert date time
+    q_obs['Date-Time'] = pd.to_datetime(q_obs['Date-Time'], format='%m/%d/%y %H:%M')
+    q_obs['Date-Time'] = q_obs['Date-Time'] - pd.Timedelta(
+        days=1)  # somehow this data was offset by one day, verifed w/ precip data
[…]
