### 2024-06-18 2766111
updated black box testing to be more portable, will take a bit more work to incoporate with github actions.
