+++ testing/black_box/big_spring/conftest.py
@@ -6,9 +6,9 @@ import getpass
 from datetime import datetime
 from concurrent.futures import ThreadPoolExecutor, as_completed
 
-extra_path = f"{os.environ['HOME']}/Documents/Repos/Forked/pytRIBS"
-if extra_path not in sys.path:
-    sys.path.append(extra_path)
+# extra_path = f"{os.environ['HOME']}/Documents/Repos/Forked/pytRIBS"
+# if extra_path not in sys.path:
+#     sys.path.append(extra_path)
 
 from pytRIBS.classes import Model as model
 from pytRIBS.classes import Results as results
