+++ testing/black_box/get_benchmarks.sh
@@ -4,7 +4,7 @@
 curl -sS -o hj "https://zenodo.org/records/10909507/files/happy_jack.gz?download=1"
 curl -sS -o bs "https://zenodo.org/records/10951574/files/big_spring.gz?download=1"
 
-path_to_dir="testing/black_box/benchmark_cases"
+path_to_dir="benchmarks"
 
 mkdir -p $path_to_dir
 
