### 2024-06-18 43f7844
Update conftest.py
Accidentally had the parallel binary provided for Happy Jack
