### 2024-06-19 7da8de8
updated how the condition where Nwt = BedrockDepth by setting the beta coefs to zero when the moisture profile matches that of the moisture profile for the above conditions. Also fixed waterbalance calcs in testing, used theta S instead of porosity following tRIBS
