+++ tests/tHydroTests/tHydroModelTests.cpp
@@ -1,0 +1,2 @@
+#include <src/tHydro/tHydroModel.h>
+#include <gtest/gtest.h>
