+++ src/utilities/RainInputCheck.cpp
@@ -30,9 +30,9 @@
 //
 //=========================================================================
 
-#include<iostream.h>
-#include<iomanip.h>
-#include<fstream.h>
+#include<iostream>
+#include<iomanip>
+#include<fstream>
 #include<stdlib.h>
 #include<stdio.h>
 #include<string.h>
