+++ src/tArray/tArray.h
@@ -27,8 +27,8 @@
 
 
 #ifdef ALPHA_64
-  #include <iostream.h>
-  #include <fstream.h>
+  #include <iostream>
+  #include <fstream>
   #include <assert.h>
 #elif defined LINUX_32
   #include <iostream>
@@ -39,12 +39,12 @@
   #include <fstream>
   #include <cassert>
 #elif defined WIN
-  #include <iostream.h>
-  #include <fstream.h>
+  #include <iostream>
+  #include <fstream>
   #include <assert.h>
 #else 
-  #include <iostream.h>
-  #include <fstream.h>
+  #include <iostream>
+  #include <fstream>
   #include <assert.h>
 #endif
 
