+++ src/Headers/globalFns.h
@@ -32,7 +32,7 @@
 #include "src/tArray/tMatrix.h"
 
 #ifdef ALPHA_64
-  #include <iostream.h>
+  #include <iostream>
   #include <math.h>
 #elif defined LINUX_32 //TODO update to LINUX
   #include <iostream>
@@ -43,10 +43,10 @@
   #include <cmath>
 
 #elif defined WIN
-  #include <iostream.h>
+  #include <iostream>
   #include <math.h>
 #else 
-  #include <iostream.h>
+  #include <iostream>
   #include <math.h>
 #endif
 
