+++ tests/tHydroTests/tHydroModelTests.cpp
@@ -1,2 +1,9 @@
-#include <src/tHydro/tHydroModel.h>
+#include "src/tHydro/tHydroModel.h"
 #include <gtest/gtest.h>
+
+TEST(HelloTest, BasicAssertions) {
+  // Expect two strings not to be equal.
+  EXPECT_STRNE("hello", "world");
+  // Expect equality.
+  EXPECT_EQ(7 * 6, 42);
+}
