+++ src/utilities/InitialGW.cpp
@@ -42,9 +42,9 @@
 **                        -------------                 ----------------- 
 ***************************************************************************/
 
-#include<iostream.h>
-#include<iomanip.h>
-#include<fstream.h>
+#include<iostream>
+#include<iomanip>
+#include<fstream>
 #include<stdlib.h>
 #include<string.h>
 #include<stdio.h>
@@ -106,57 +106,57 @@ int main(int argc, char *argv[]){
 	double cellsize, nodata;
 	char *names[5];
 	
-	cout<<"\n-----------------------------------------------------------------"
+	std::cout<<"\n-----------------------------------------------------------------"
 		<<"-------";
-	cout<<"\n\n\t\t tRIBS -- Version 1.0";
-	cout<<"\n\t\t tRIBS Model: InitialGW Utility";
-	cout<<"\n\t\t Ralph M. Parsons Laboratory";
-	cout<<"\n\t\t Massachusetts Institute of Technology";
-	cout<<"\n\n\t\t Release, 9/2001 \n\n";
-	cout<<"-----------------------------------------------------------------"
-		<<"-------"<<endl;
+	std::cout<<"\n\n\t\t tRIBS -- Version 1.0";
+	std::cout<<"\n\t\t tRIBS Model: InitialGW Utility";
+	std::cout<<"\n\t\t Ralph M. Parsons Laboratory";
+	std::cout<<"\n\t\t Massachusetts Institute of Technology";
+	std::cout<<"\n\n\t\t Release, 9/2001 \n\n";
+	std::cout<<"-----------------------------------------------------------------"
+		<<"-------"<<std::endl;
 	
-	cout<<"\nThis program implements a TOPMODEL approach of defining a steady-state GW\n"
+	std::cout<<"\nThis program implements a TOPMODEL approach of defining a steady-state GW\n"
 		<<"surface based on the contributing area distribution in the basin and the\n"
-		<<"local slope of the terrain surface."<<endl;
+		<<"local slope of the terrain surface."<<std::endl;
 	
 	// Command line arguments
 	if(argc != 2) {
-		cout<<"\nUsage: "<< argv[0] << "  *.gw\n" << endl;
-		cout<<"Order of names in the file: " << endl;
-		cout<<"\t 1. DEM file (*.asc)\n";
-		cout<<"\t 2. Flow Accumulations file (*.asc)\n";
-		cout<<"\t 3. Flow Directions file (*.asc)\n";
-		cout<<"\t 4. Soil reclassification table (*.sdtt)\n";
-		cout<<"\t 5. Basin area (km^2)\n";
-		cout<<"\t 6. Baseflow of Reference Q(0) at outlet (m^3/sec)\n";
-		cout<<"\t 7. Zero Baseflow Q_0 from recession analysis (m^3/sec)\n";
-		cout<<"\t 8. Threshold value for stream network (# of pixels)\n";
-		cout<<"\t 9. Size of MA smoothing window (3 or 5 pixels)\n";
-		cout<<"\t 10. Number of smoothing iterations (2 or greater)\n";
-		cout<<"\t 11. Average basin GWT depth (~ -ln(Q(0)/Q_0)/f)\n\n";
[…]
