### 2024-08-01 #6747 @alessandroamaranto
I am happy with what I see, I am actually going to use it myself :).
I noticed that the manuscript doesn't include a 'state of the field' section. However I believe this may not be critical
in light of the fact that tRIBS serves as an open-source counterpart to a well-established software, and that the 'statement of needs' section already touches on relevant aspects.
