### 2024-08-07 #6747 @gutabeshu
The installation procedures are clear. The model runs successfully, and the output result (Figure 3) is reproducible. I have minor comment regarding Figure 1of the manuscript, under "Forcing and Parameters Inputs": It seems redundant to list both "Meteorological Stations" and "Precipitation Stations". I would expect "Meteorological Stations" to suffice unless the two terms refer to different types of stations. Please clarify if they represent distinct entities.
