+++ tests/tHydroTests/tHydroModelTests.cpp
@@ -1,6 +1,5 @@
 #include "src/tHydro/tHydroModel.h"
-#include <gtest/gtest.h>
-
+#include "gtest/gtest.h"
 TEST(HelloTest, BasicAssertions) {
   // Expect two strings not to be equal.
   EXPECT_STRNE("hello", "world");
