+++ CMakeLists.txt
@@ -309,9 +309,20 @@ endif ()
 
 enable_testing ()
 
+include(FetchContent)
+FetchContent_Declare(
+  googletest
+  # Specify the commit you depend on and update it regularly.
+  URL https://github.com/google/googletest/archive/5376968f6948923e2411081fd9372e71a59d8e77.zip
+)
+# For Windows: Prevent overriding the parent project's compiler/linker settings
+set(gtest_force_shared_crt ON CACHE BOOL "" FORCE)
+FetchContent_MakeAvailable(googletest)
+
 add_executable(
   tHydroTests
   ${PROJECT_SOURCE_DIR}/tests/tHydrotests/THydroModelTests.cpp
+  ${PROJECT_SOURCE_DIR}/src/tHydro/tHydroModel.h
 )
 target_link_libraries(
   tHydroTests
