### 2024-08-07 #6747 @WrenRaming
Hi @gutabeshu, thank you for your feedback! Though it seems redundant to list both "Meteorological Stations" and "Precipitation Stations", the way tRIBS is formatted it ingests precipitation separately from meteorological forcings such as wind speed, etc. To that end the redundancy reflects how these data are typically ingested by tRIBS.
