### 2024-08-14 #6747 @cheginit
@WrenRaming I am doing my final checks before recommending acceptance. Please address the following in the paper:

- [ ] L17: Replace "software" with "software package".
- [ ] If C. J. Cederstrom and A. Ko have ORCIDs please add them.
- [ ] Use a consistent style for paper titles in References section; either use title style (upper case for the first letter of all the words) like the first item or only making the first letter upper case like the second item.

Once you made these changes, run `@editorialbot generate pdf` command to check the changes, then carry on to doing the tasks listed in "Additional Author Tasks After Review is Complete" in the previous comment.
