### 2024-08-15 #6747 @WrenRaming
Hi @cheginit I have addressed the following:

> * [x]  L17: Replace "software" with "software package".
> * [x]  If C. J. Cederstrom and A. Ko have ORCIDs please add them. (Only updated for Cederstrom)
> * [x]  Use a consistent style for paper titles in References section; either use title style (upper case for the first letter of all the words) like the first item or only making the first letter upper case like the second item.

I have also archived the software package w/ Zenodo, here is the DOI: [10.5281/zenodo.13324178](https://zenodo.org/doi/10.5281/zenodo.13324178). The version number has gone unchanged (5.2.0) since no changes were required. I double checked the author list, title, and license between the archive and the paper and they both match.
