### 2024-08-16 #6747 @cheginit
@WrenRaming The version of the archive should match the software version, i.e., 5.2. Please change the Zenodo version to 5.2. Also, if the version is supposed to be 5.2.0 as opposed to 5.2., do so.
