### 2024-08-19 #6747 @cheginit
@WrenRaming While doing my final checks, I realized that there are no end-to-end examples in the documentation. This can make it hard for users to start working with the software and even figure out the capabilities of the software at a glance. A quick-guide style example that doesn't delve deep into the complexities of the software yet demonstrates how a user can start running the model and plot the results, is missing.

Here's my suggestion: Have a quick guide section in the documentation that starts from scratch, i.e., installation commands for installing prerequisites (e.g., `brew install openmpi cmake` for macOS or `sudo apt install openmpi-bin openmpi-dev openmpi-common openmpi-doc libopenmpi-dev cmake` for ubuntu), then building the package from source (the `Compilation Instructions`), then running the model with an example input file (for reproducing the benchmark cases), and finally plotting the results using Python packages or anything that you would normally use. Essentially, the purpose is for a user to be able to reproduce the benchmark examples and use it as a reference to run their own cases.

Just to demonstrate my point, I couldn't get the benchmark cases to run even when I successfully built tRIBS from source, since the run instructions were not clear to me. I didn't know what to do with the extracted benchmark files. The readme on Zenodo says there's an HTML readme file, but when I extracted the file, there's just one big file. I tried running it with `mpirun` and even passing it as an input to `tRIBSPar`
