### 2024-08-20 #6747 @WrenRaming
Hi @cheginit, for the benchmarks the readme file is in the zip.  I do understand your point about the value of the quick start guide. I'll work on that asap.
