### 2024-08-22 #6747 @WrenRaming
@cheginit  I have added a Quick Start Guide to the documentation. You can find it [here](https://tribshms.readthedocs.io/en/latest/man/QuickStart.html). Let me know if this is what you were looking for and if you have any additional suggestions for improvement.
