### 2024-08-23 #6747 @cheginit
@WrenRaming It looks good to me. I figured out the issue that I had, when I downloaded the `big_spring.gz` file and extracted it, the extracted file didn't have any extension, so I had to manually rename it to `big_spring.zip` and unzip it again to see the whole folder with all the files. I think it'd better to add a note that if such a thing happens, users know what to do. Here's how it looks like on my machine after manually renaming the extracted file and extracting it again:

I was able to run the case on my M2 Mac Studio with `-n 8` in 6 min and reproduce the results.
