### 2024-08-23 #6747 @WrenRaming
@cheginit I have updated both Zenodo benchmark repos notifying potential users of this issue. Also I have updated  the quick start guide to include: brew install openmpi , not open-mpi.
