### 2024-09-10 #6747 @kthyng
Thank you for the ping @WrenRaming. I believe that despite the unchecked boxes, @alessandroamaranto has stated satisfaction here: https://github.com/openjournals/joss-reviews/issues/6747#issuecomment-2262712545 so we will proceed.
